code cleanup to permit building with C++.
Replace assert() and interr() arguments with enum values.

Modify symini.cpp to emit type-correct initializations.

This cleans up tons of type errors.

Fix type issues with second argument to error().

Fix for dlen(). Despite the comments, the two versions of dlen()
were, in fact, not the same.

More propagation of SPTR types.

Cleanup warnings in ccffinfo.c

Fix some error calls

Get bihutil.c, expreg.c, and xref.c to compile clean as C++

More C++ changes

- get llmputil.c to compile clean as C++
- fix up a number of typing issues in mwd.c

Convert llutil.c to C++

Convert llmputil.c to C++

Convert llsched.c
Convert symacc.c
Convert rmsmove.c

+++ tools/flang2/flang2exe/expreg.c
@@ -337,16 +337,16 @@ assign1(int rtype)
            * the arg register that's assigned to its address as the
            * address expression in the the load.
            */
-          addr = ad1ili(IL_ARDF, (int)ADDRESSG(sym));
+          addr = ad1ili(IL_ARDF, ADDRESSG(sym));
           if (flg.endian &&
               (DTYPEG(sym) == DT_INT8 || DTYPEG(sym) == DT_LOG8)) {
             if (!XBIT(124, 0x400))
               /* 32bits of significance in 64 bits */
-              addr = ad3ili(IL_AADD, addr, ad_aconi((INT)4), 0);
+              addr = ad3ili(IL_AADD, addr, ad_aconi(4), 0);
           }
           RAT_ADDR(rat) = addr; /* switch addr expr */
-          ilix = ad3ili((int)ILI_OPC(ilix), addr, (int)ILI_OPND(ilix, 2),
-                        (int)ILI_OPND(ilix, 3));
+          ilix = ad3ili(ILI_OPC(ilix), addr, ILI_OPND(ilix, 2),
+                        ILI_OPND(ilix, 3));
           ilt = addilt(ilt, ilitmp = ad2ili(MV_RTYPE(rtype), ilix, areg));
           ADDNODE(list, ilitmp);
         }
